new deployment script for rain which makes use of the NFS server and does not transfer 55 MByte through the thrift interface!

+++ RelayControl/drones/rain_deploy/main.sh
@@ -6,8 +6,18 @@ RAIN=/opt/rain
 mkdir $RAIN
 rm -rf $RAIN/*
 
+# Files are in the NFS share
+if [ -d /mnt/share ]; then
+echo "ATTENTION: mounting /mnt/share"
+mount -t nfs monitor0:/mnt/arr0/share /mnt/share
+
 # Update the files
-cp driver.zip $RAIN
+cp /mnt/share/packages/rain/driver.zip $RAIN
 cd $RAIN
 unzip driver.zip
 rm driver.zip
+
+echo "umount /mnt/share"
+umount /mnt/share
+
+fi
